Use base train limits
Found a way to use the base train limits when dispatching.
Using the LTN limit signal is priority. If not given, check if a vanilla train limit is set, subtract the amount of vanilla trains reserved for that station and treat the result as train limit.
Of course the dispatcher won't send a train to this delivery even if the vanilla train would have left the requester long time before the LTN train arrives.
But I think it's a good workaround to make use of the base train limits.